refactor: move TermsContractParameters namespace into its own module
