+++ src/adapters/collateralized_simple_interest_loan_adapter.ts
@@ -3,6 +3,7 @@ import { ContractsAPI } from "src/apis";
 import { Assertions } from "src/invariants";
 import { BigNumber } from "utils/bignumber";
 import * as singleLineString from "single-line-string";
+import { TermsContractParameters } from "./terms_contract_parameters";
 
 const MAX_COLLATERAL_TOKEN_INDEX_HEX = TermsContractParameters.generateHexValueOfLength(2);
 const MAX_COLLATERAL_AMOUNT_HEX = TermsContractParameters.generateHexValueOfLength(23);
@@ -12,18 +13,6 @@ export interface CollateralizedTermsContractParameters {
     collateralTokenIndex: BigNumber;
     collateralAmount: BigNumber;
     gracePeriodInDays: BigNumber;
-}
-
-export namespace TermsContractParameters {
-    export function generateHexValueOfLength(length: number): string {
-        return "0x" + "f".repeat(length);
-    }
-
-    export function bitShiftLeft(target: BigNumber, numPlaces: number): BigNumber {
-        const binaryTargetString = target.toString(2);
-        const binaryTargetStringShifted = binaryTargetString + "0".repeat(numPlaces);
-        return new BigNumber(binaryTargetStringShifted, 2);
-    }
 }
 
 export const CollateralizedAdapterErrors = {
